Правки по финальному ревью - раскидал всё по файликам. - вынес обработку следующей загадки в отдельную функцию - использую в Alert weak self - extension для GameState для упрощения кода - избавился от -1 у currentRiddleIndex и добавил assert - добавил имена переменным в enum - переиспользую NumberFormatter
