+++ MovieQuiz/IMDBGateway/IMDBMovie.swift
@@ -1,0 +1,4 @@
+struct IMDBMovie {
+    let name: String
+    let rating: Double
+}
