+++ MovieQuiz/MovieQuizModel/Answer.swift
@@ -1,0 +1,4 @@
+enum Answer {
+    case yes
+    case no
+}
