+++ MovieQuiz/MovieQuizModel/GameState.swift
@@ -1,0 +1,6 @@
+enum GameState {
+    case nextRiddle(riddle: MovieRiddle, riddleNum: Int, riddlesCount: Int)
+    case positiveAnswer
+    case negativeAnswer
+    case gameEnded(correctAnswers: Int, riddlesCount: Int)
+}
