+++ MovieQuiz/MovieQuizModel/RiddleSign.swift
@@ -1,0 +1,4 @@
+enum RiddleSign: String {
+    case more = "больше"
+    case less = "меньше"
+}
